requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-ckeditor==6.7.0
